requirements-dev.txt
coverage
syrupy
pytest
tenacity
pydantic_requests
moviepy